feat: pass atoms dataframe to C++ via numpy
The previous commit introduced code to pass atoms to C++ via Arrow
tables.  This would've been a better solution, because it would never
require copying data.  Unfortunately, it requires linking to the arrow
and pyarrow libraries, and this turned out to be prohibitively
difficult.  I did get it to work on my laptop, but (i) it didn't work on
the cluster due to CXX11 ABI mismatches and (ii) it made it impossible
to distribute binary wheels.

Passing atom data with numpy arrays does often require copying.  The
reason is that the initial filtering step often causes the atoms
dataframe to be split into multiple chunks.  I assume this is a
side-effect of some sort of parallel processing.  Even so, the numpy
approach should still be better than having the whole loop in python,
which is what I had before.  For one thing, this should cause many fewer
heap allocations.